requirements.txt
python-dotenv
groq
AppOpener
pywhatkit
bs4
pillow
rich
requests
keyboard
cohere
googlesearch-python
selenium
mtranslate
pygame
edge-tts
PyQt5
webdriver-manager
gtts
Flask
google-generativeai
google-genai
googletrans==4.0.0-rc1
pyautogui
opencv-python
unidecode
pvporcupine
pyaudio
dateparser